# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Biblioteca",
        "description": "Sistema de Gerenciamento de Biblioteca",
        "contact": {
            "name": "Equipe de Suporte",
            "email": "suporte@biblioteca.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Servidor de API da Biblioteca"
        }
    ],
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Lista todos os usuários",
                "operationId": "3ab81e3f9b24b4f36ab08d72349ac632",
                "responses": {
                    "200": {
                        "description": "Lista de usuários",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Criar novo usuário",
                "operationId": "ba548b97148bb5c0d8cab3bc3f128bfd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Obter detalhes de um usuário específico",
                "operationId": "daaf7a6169ee9348c3fddb07778598ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Atualizar usuário existente",
                "operationId": "59e98fba99e974815e27979dd58e2175",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Remover usuário",
                "operationId": "c30c64b8b7730b65c7714c9af4a6f5fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário removido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            }
        },
        "/api/v1/users/{id}/loans": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Obter empréstimos do usuário",
                "operationId": "911d1da132e024088323104de305a3e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empréstimos do usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Loan"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Loan": {
                "title": "Empréstimo",
                "description": "Modelo de empréstimo de livro",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "book_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "delayed",
                            "returned"
                        ],
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "Usuário",
                "description": "Modelo de usuário da biblioteca",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Luis Henrique"
                    },
                    "email": {
                        "type": "string",
                        "example": "luis@example.com"
                    },
                    "registration_number": {
                        "type": "string",
                        "example": "REG123"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "UserCreateRequest": {
                "title": "Requisição de Criação de Usuário",
                "required": [
                    "name",
                    "email",
                    "registration_number",
                    "password"
                ],
                "properties": {
                    "name": {
                        "description": "Nome completo do usuário",
                        "type": "string",
                        "example": "Luis Henrique"
                    },
                    "email": {
                        "description": "Endereço de email do usuário",
                        "type": "string",
                        "format": "email",
                        "example": "luis@example.com"
                    },
                    "registration_number": {
                        "description": "Número de registro único do usuário",
                        "type": "string",
                        "example": "REG12345"
                    },
                    "password": {
                        "description": "Senha do usuário",
                        "type": "string",
                        "format": "password",
                        "example": "senha123"
                    }
                },
                "type": "object"
            },
            "UserUpdateRequest": {
                "title": "Requisição de Atualização de Usuário",
                "properties": {
                    "name": {
                        "description": "Nome completo do usuário",
                        "type": "string",
                        "example": "Luis Henrique"
                    },
                    "email": {
                        "description": "Endereço de email do usuário",
                        "type": "string",
                        "format": "email",
                        "example": "luis@example.com"
                    },
                    "registration_number": {
                        "description": "Número de registro único do usuário",
                        "type": "string",
                        "example": "REG12345"
                    },
                    "password": {
                        "description": "Nova senha do usuário (opcional)",
                        "type": "string",
                        "format": "password",
                        "example": "novaSenha123"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}